# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ATLAS BI - Commercial Dashboard API for NIADCON
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List all sales
      parameters:
        - name: consultantId
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
          description: "Month number 1-12"
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Register a new sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a sale by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    patch:
      operationId: updateSale
      tags: [sales]
      summary: Update a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List commercial goals
      parameters:
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create or update a commercial goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        consultantId:
          type: integer
        consultantName:
          type: ["string", "null"]
        product:
          type: string
        amount:
          type: number
        quantity:
          type: integer
        saleDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - consultantId
        - product
        - amount
        - quantity
        - saleDate
        - createdAt
    SaleInput:
      type: object
      properties:
        consultantId:
          type: integer
        product:
          type: string
          minLength: 1
        amount:
          type: number
        quantity:
          type: integer
        saleDate:
          type: string
          format: date
        notes:
          type: string
      required:
        - consultantId
        - product
        - amount
        - quantity
        - saleDate
    SaleUpdate:
      type: object
      properties:
        consultantId:
          type: integer
        product:
          type: string
        amount:
          type: number
        quantity:
          type: integer
        saleDate:
          type: string
          format: date
        notes:
          type: string
    Goal:
      type: object
      properties:
        id:
          type: integer
        consultantId:
          type: ["integer", "null"]
        consultantName:
          type: ["string", "null"]
        month:
          type: integer
        year:
          type: integer
        targetAmount:
          type: number
        targetQuantity:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - month
        - year
        - targetAmount
        - createdAt
    GoalInput:
      type: object
      properties:
        consultantId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        targetAmount:
          type: number
        targetQuantity:
          type: integer
        description:
          type: string
      required:
        - month
        - year
        - targetAmount
    GoalUpdate:
      type: object
      properties:
        targetAmount:
          type: number
        targetQuantity:
          type: integer
        description:
          type: string
...