# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ATLAS BI - Commercial Dashboard API for NIADCON
 * OpenAPI spec version: 0.1.0
 */

export interface GoalInput {
  consultantId?: number;
  month: number;
  year: number;
  targetAmount: number;
  targetQuantity?: number;
  description?: string;
}