# lib/api-zod/src/generated/types/saleUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ATLAS BI - Commercial Dashboard API for NIADCON
 * OpenAPI spec version: 0.1.0
 */

export interface SaleUpdate {
  consultantId?: number;
  product?: string;
  amount?: number;
  quantity?: number;
  saleDate?: Date;
  notes?: string;
}